result/pyproject.toml
[project]
name = "result"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agent>=0.1.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
